ai-engine/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
paddlepaddle==2.6.1
paddleocr==2.7.0.3
pillow==10.3.0
